Options for the src, test, and dependency paths

+++ watch.js
@@ -2,37 +2,68 @@ var fs = require("fs");
 var path = require("path");
 var child = require("child_process");
 var log = require("./log");
-var Minimatch = require("minimatch").Minimatch;
+var files = require("./files");
+var index = require("./index");
+var minimatch = require("minimatch");
+var Watchpack = require("watchpack");
 
 function stripWatchArg(arg) {
   return arg !== "-w" && arg !== "--watch";
 }
 
-var match = new Minimatch("{src,test,bower_components}/**/*");
-
-function watch(match, act) {
-  require("watch").watchTree(".", {
-    interval: 1337,
-    ignoreDotFiles: true
-  }, function(f, curr, prev) {
-    if (!(typeof f === "object" && prev === null && curr === null)) {
-      if (match.match(f)) {
-        act(f);
-      }
-    }
+function topLevel(directories) {
+  return directories.map(function(a){
+    return a.split(path.sep)[0];
   });
 }
 
+function watch(directories, act) {
+  var watchpack = new Watchpack();
+
+  watchpack.watch([], directories, Date.now() - 10000);
+
+  watchpack.on("change", act);
+}
+
 module.exports = function() {
-  var mod = path.join(__dirname, "/index"),
-      args = process.argv.filter(stripWatchArg).slice(2);
+  var mod = path.join(__dirname, "/index");
+
+  var args = process.argv.filter(stripWatchArg).slice(2);
+
   var p = child.fork(mod, args);
-  var change = function() {
-    p.kill("SIGTERM");
-    log("Source tree changed; restarting:");
-    p = child.fork(mod, args);
-  };
-  watch(match, change);
+
+  var srcPath = topLevel(index.opts.opts.srcPath);
+
+  var testPath = topLevel(index.opts.opts.testPath);
+
+  var dependencyPath = topLevel(index.opts.opts.dependencyPath);
+
+  var directories = srcPath.concat(testPath).concat(dependencyPath);
+
+  var localGlobs = files.localGlobs(index.opts.opts);
+
+  var testGlobs = files.testGlobs(index.opts.opts);
+
+  var dependencyGlobs = files.dependencyGlobs(index.opts.opts);
+
+  var globs = localGlobs.union(testGlobs).union(dependencyGlobs)
+
+  var sourceGlobs = globs.sources();
+
+  var ffiGlobs = globs.ffis();
+
+  var fileGlobs = sourceGlobs.concat(ffiGlobs);
+
+  watch(directories, function(change){
+    var matched = fileGlobs.some(function(a){
+      return minimatch(change, a);
+    });
+    if (matched) {
+      p.kill("SIGTERM");
+      log("Source tree changed; restarting:");
+      p = child.fork(mod, args);
+    }
+  });
 };
 
 module.exports.watch = watch;
